K Nearest Neighbor Algorithm (KNN)
This program was executed on the file 'glass.txt'. Before starting the algorithm, the data was divided into training and test sets. 75% of the data was taken as training data and the rest 25% as test. Random was used to pick the training data. After the data selection is made, we start the classification process.
The Classification method used here was the KNN classification fro unknown records. Implemented Using Majority class model and the Distance-weighted model.

Following is how the program works:

User is given a limited choice for the neighbors K. User is asked to enter value in (1, 5, 10)
User compiles/ executes the program, user is asked to choose a value of 'K' from the list provided. If user enters a wrong choice of 'K', user is asked if he would like to continue or exit. If he chooses exit, the program ends. But, if the user want's to continue, user is prompted again with the dialog to enter a value of 'K' .
After the user has entered the needed value of K, the algorithm works with this value of 'K' using the above two models "Majority class", and "Distance-Weighted".
The output is displayed on the screen. First the result for the Majority class is display with the ["TestData Identifier", "Class(classified to)"] along with the accuracy of the classification. After user reviewing this result, user hits ok. Now the result for the Distance-weighted model is displayed with the result using the same format of the Majority class.
After user review, user is asked if he would like to test using different values of 'K'. If he chooses to then the process repeats. If user chooses no, then user is given a thank you message.